requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bpemb==0.3.4
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.1
colorlover==0.3.0
comm==0.1.2
conllu==4.5.2
contourpy==1.0.7
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
Deprecated==1.2.13
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
flair==0.11.3
fonttools==4.38.0
fst-pso==1.8.1
ftfy==6.1.1
future==0.18.3
FuzzyTM==2.0.5
gdown==4.4.0
gensim==4.3.0
huggingface-hub==0.12.0
hyperopt==0.2.7
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.1
ipython==8.9.0
ipywidgets==8.0.4
Janome==0.4.2
jedi==0.18.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kaleido==0.2.1
kiwisolver==1.4.4
konoha==4.6.5
langdetect==1.0.9
lxml==4.9.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
miniful==0.0.6
more-itertools==9.0.0
mpld3==0.3
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
overrides==3.1.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.13.0
pptree==3.1
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyFUME==0.2.25
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
segtok==1.5.11
sentencepiece==0.1.95
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
sqlitedict==2.1.0
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.1.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.0
tweet-preprocessor==0.6.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
widgetsnbextension==4.0.5
Wikipedia-API==0.5.8
wrapt==1.14.1
zipp==3.12.0
